chore: modify directory structure
