+++ app/common/utils/index.ts
@@ -1,1 +1,15 @@
-export {};
+import probeImageSize from 'probe-image-size';
+
+export const getImageSize = async (
+  url: string
+): Promise<{ width?: number; height?: number }> => {
+  try {
+    const result = await probeImageSize(url);
+    return {
+      height: result.height,
+      width: result.width,
+    };
+  } catch (e) {
+    return {};
+  }
+};
